Refactor StableVault page and misc changes
