+++ components/Icons/StableVaultIcon/index.js
@@ -1,0 +1,33 @@
+
+import { memo } from 'react'
+import SvgIcon from '@material-ui/core/SvgIcon'
+import { makeStyles } from '@material-ui/core/styles'
+import clsx from 'clsx'
+
+const useStyles = makeStyles(() => ({
+  root: {
+    width: 20,
+    height: 20
+  }
+}));
+
+const StableVaultIcon = ({
+  className,
+  viewBox,
+  ...rest
+}) => {
+  const classes = useStyles();
+
+  return (
+    <SvgIcon viewBox={viewBox || '0 0 24 24'} {...rest} className={clsx(classes.root, className)}>
+		<path d="M8.7999 7.1C9.8999 6.5 10.8999 5.9 11.8999 5.3C11.9999 5.2 12.0999 5.3 12.0999 5.3C14.3999 6.6 16.6999 7.9 18.9999 9.3C19.3999 9.5 19.4999 9.9 19.4999 10.3C19.4999 11.2 19.4999 12.1 19.4999 13C19.4999 13.5 19.1999 13.9 18.6999 14.1C17.9999 14.6 16.9999 15 16.2999 15.5C14.8999 16.3 13.5999 17.1 12.1999 17.8C11.9999 17.9 11.8999 17.9 11.6999 17.8C9.4999 16.6 7.3999 15.3 5.1999 14.1C4.5999 13.8 4.3999 13.3 4.3999 12.7C4.3999 11.8 4.3999 10.9 4.3999 10C4.6999 10.1 4.8999 10.2 5.0999 10.4C7.1999 11.6 9.3999 12.8 11.4999 14.1C11.7999 14.3 12.0999 14.3 12.3999 14.1C13.4999 13.4 14.6999 12.8 15.7999 12.1C15.8999 12 15.9999 12 16.0999 11.9C16.2999 11.8 16.2999 11.6 16.0999 11.4C15.9999 11.3 15.7999 11.2 15.5999 11.1C13.4999 9.9 11.3999 8.7 9.2999 7.5C9.1999 7.4 8.9999 7.3 8.7999 7.1Z" fill="#6E6B7B"/>
+		<path d="M19.5 8.79998C19.2 8.69998 18.9 8.49998 18.7 8.39998C16.7 7.19998 14.7 6.09998 12.7 4.89998C12.2 4.59998 11.7 4.59998 11.2 4.89998C10.2 5.39998 9.30005 5.99998 8.30005 6.49998C8.00005 6.69998 7.60005 6.79998 7.60005 7.09998C7.60005 7.39998 8.00005 7.49998 8.30005 7.69998C10.3 8.89998 12.3 9.99998 14.3 11.1C14.6 11.2 14.8 11.4 15.1 11.6C15.1 11.7 15 11.7 15 11.8C14.1 12.3 13.1 12.9 12.2 13.4C12.1 13.5 12 13.5 11.8 13.4C9.30005 12 6.90005 10.7 4.50005 9.29998C4.40005 9.19998 4.30005 9.09998 4.30005 8.99998C4.30005 7.99998 4.30005 6.99998 4.30005 5.99998C4.30005 5.39998 4.50005 4.99998 5.10005 4.69998C6.60005 3.89998 8.00005 2.99998 9.50005 2.19998C10.1 1.89998 10.6 1.59998 11.2 1.19998C11.7 0.899976 12.2 0.899976 12.7 1.19998C14.7 2.39998 16.7 3.49998 18.7 4.69998C19.2 4.99998 19.5 5.39998 19.5 6.09998C19.4 6.89998 19.5 7.79998 19.5 8.79998Z" fill="#6E6B7B"/>
+		<path d="M19.4999 14.2C19.4999 14.3 19.4999 14.4 19.4999 14.5C19.4999 15.5 19.4999 16.4 19.4999 17.4C19.4999 18 19.2999 18.3 18.7999 18.6C16.7999 19.8 14.7999 20.9 12.7999 22.1C12.2999 22.4 11.7999 22.4 11.1999 22.1C9.1999 20.9 7.1999 19.8 5.1999 18.6C4.6999 18.3 4.3999 17.9 4.3999 17.3C4.3999 16.4 4.3999 15.4 4.3999 14.5C4.6999 14.6 4.8999 14.7 5.0999 14.9C7.1999 16.1 9.2999 17.3 11.3999 18.6C11.6999 18.8 11.9999 18.8 12.3999 18.6C14.4999 17.4 16.5999 16.2 18.6999 15C19.0999 14.8 19.2999 14.5 19.4999 14.2Z" fill="#6E6B7B"/>
+		<path d="M8.80005 7.1C9.90005 6.5 10.9 5.9 11.9 5.3C12 5.2 12.1 5.3 12.1 5.3C14.4 6.6 16.7 7.9 19 9.3C19.4 9.5 19.5 9.9 19.5 10.3C19.5 11.2 19.5 12.1 19.5 13C19.5 13.5 19.2 13.9 18.7 14.1C17.9 14.6 17 15 16.2 15.5C14.8 16.3 13.5 17.1 12.1 17.8C11.9 17.9 11.8 17.9 11.6 17.8C9.40005 16.6 7.30005 15.3 5.10005 14.1C4.50005 13.8 4.30005 13.3 4.30005 12.7C4.30005 11.8 4.30005 10.9 4.30005 10C4.60005 10.1 4.80005 10.2 5.00005 10.4C7.10005 11.6 9.30005 12.8 11.4 14.1C11.7 14.3 12 14.3 12.3 14.1C13.4 13.4 14.6 12.8 15.7 12.1C15.8 12 15.9 12 16 11.9C16.2 11.8 16.2 11.6 16 11.4C15.9 11.3 15.7 11.2 15.5 11.1C13.4 9.9 11.3 8.7 9.20005 7.5C9.20005 7.4 9.00005 7.3 8.80005 7.1Z" fill="#6E6B7B" stroke="white" strokeMiterlimit="10"/>
+		<path d="M19.5 8.79998C19.2 8.69998 18.9 8.49998 18.7 8.39998C16.7 7.19998 14.7 6.09998 12.7 4.89998C12.2 4.59998 11.7 4.59998 11.2 4.89998C10.2 5.39998 9.30005 5.99998 8.30005 6.49998C8.00005 6.69998 7.60005 6.79998 7.60005 7.09998C7.60005 7.39998 8.00005 7.49998 8.30005 7.69998C10.3 8.89998 12.3 9.99998 14.3 11.1C14.6 11.2 14.8 11.4 15.1 11.6C15.1 11.7 15 11.7 15 11.8C14.1 12.3 13.1 12.9 12.2 13.4C12.1 13.5 12 13.5 11.8 13.4C9.30005 12 6.90005 10.7 4.50005 9.29998C4.40005 9.19998 4.30005 9.09998 4.30005 8.99998C4.30005 7.99998 4.30005 6.99998 4.30005 5.99998C4.30005 5.39998 4.50005 4.99998 5.10005 4.69998C6.60005 3.89998 8.00005 2.99998 9.50005 2.19998C10.1 1.89998 10.6 1.59998 11.2 1.19998C11.7 0.899976 12.2 0.899976 12.7 1.19998C14.7 2.39998 16.7 3.49998 18.7 4.69998C19.2 4.99998 19.5 5.39998 19.5 6.09998C19.4 6.89998 19.5 7.79998 19.5 8.79998Z" fill="#6E6B7B" stroke="white" strokeMiterlimit="10"/>
+		<path d="M19.4 14.2C19.4 14.3 19.4 14.4 19.4 14.5C19.4 15.5 19.4 16.4 19.4 17.4C19.4 18 19.2 18.3 18.7 18.6C16.7 19.8 14.7 20.9 12.7 22.1C12.2 22.4 11.7 22.4 11.1 22.1C9.10005 20.9 7.10005 19.8 5.10005 18.6C4.60005 18.3 4.30005 17.9 4.30005 17.3C4.30005 16.4 4.30005 15.4 4.30005 14.5C4.60005 14.6 4.80005 14.7 5.00005 14.9C7.10005 16.1 9.20005 17.3 11.3 18.6C11.6 18.8 11.9 18.8 12.3 18.6C14.4 17.4 16.5 16.2 18.6 15C19 14.8 19.2 14.5 19.4 14.2Z" fill="#6E6B7B" stroke="white" strokeMiterlimit="10"/>
+    </SvgIcon>
+  )
+}
+
+export default memo(StableVaultIcon);
