+++ utils/constants/lp-icons.js
@@ -1,27 +1,4 @@
 const LP_ICONS = Object.freeze({
-  png: '/assets/images/lp-icons/PNG.png',
-  wavax: '/assets/images/lp-icons/AVAX.png',
-  snowball: '/assets/images/lp-icons/SNOWBALL.png',
-  ether: '/assets/images/lp-icons/ETH.png',
-  eth: '/assets/images/lp-icons/ETH.png',
-  link: '/assets/images/lp-icons/LINK.png',
-  sushi: '/assets/images/lp-icons/SUSHI.png',
-  wbtc: '/assets/images/lp-icons/WBTC.png',
-  usdt: '/assets/images/lp-icons/USDT.png',
-  dai: '/assets/images/lp-icons/DAI.png',
-  aave: '/assets/images/lp-icons/AAVE.png',
-  uni: '/assets/images/lp-icons/UNI.png',
-  yfi: '/assets/images/lp-icons/YFI.png',
-  frax: '/assets/images/lp-icons/FRAX.png',
-  tusd: '/assets/images/lp-icons/TUSD.png',
-  busd: '/assets/images/lp-icons/BUSD.png',
-  s3d: '/assets/images/lp-icons/S3D.png',
-  s3f: '/assets/images/lp-icons/S3F.png',
-  vso: '/assets/images/lp-icons/VSO.png',
-  spore: '/assets/images/lp-icons/SPORE.png',
-  joe: '/assets/images/lp-icons/JOE.png',
-
-  // UPDATE ICON CONSTANTS
   PNG: '/assets/images/lp-icons/PNG.png',
   WAVAX: '/assets/images/lp-icons/AVAX.png',
   SNOB: '/assets/images/lp-icons/SNOWBALL.png',
